SucceededAssetsTests - OK

+++ AssetTool/Service/TransferReader.cs
@@ -31,7 +31,7 @@ namespace AssetTool
         public override void Move(ref ulong value) => reader.Read(ref value);
         public override void Move(ref float value) => reader.Read(ref value);
         public override void Move(ref double value) => reader.Read(ref value);
-        public override void MoveSingleOrDouble(ref double value) => value = Supports.LARGE_WORLD_COORDINATES ? reader.ReadDouble() : (double)reader.ReadSingle();
+        public override void MoveSingleOrDouble(ref double value) => value = Supports.LARGE_WORLD_COORDINATES ? reader.ReadDouble() : reader.ReadSingle();
         #endregion
 
         #region
